refactor: modularize advance request form components
- Introduced new shared components for advance information, installment fields, signature section, and read-only display to enhance code organization and reusability.
- Streamlined the advance request form logic by separating concerns into dedicated files, improving maintainability and readability.
- Updated form validation to ensure proper handling of amounts and installment data.
- Implemented a new PDF generation process after creating advance requests to improve user experience.
- Removed deprecated imports and unused code to clean up the codebase.
